thinkcoders-main/backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-session
python-dotenv
psycopg2-binary
cohere